deploy/pyproject.toml
[tool.poetry]
name = "deploy"
version = "0.1.0"
description = ""
authors = ["Merwane Hamadi <merwanehamadi@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
aws-cdk-lib = "^2.151.0"
constructs = "^10.3.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
